0114
0114

+++ back/src/DataBase/modifydata.js
@@ -30,49 +30,52 @@ module.exports = {
                 console.log('difference : ', difference);
                 var deldata = `(${difference})`
                 console.log(deldata)
-                var middlegoaldeldata =``
+                var middlegoaldeldata = ``
                 var shortgoaldeldata = ``
                 var deltable = ``
-                conn.queryAsync(`SELECT * FROM longgoal JOIN middlegoal on longgoal.longgoal_id = middlegoal.longgoal_id;`)
-                            .then((result)=>{
-                                console.log(9999999999999)
-                                console.log(result)
-                                var dellong = new Array();
-                                for(i in result){
-                                    if(difference.includes(result[i].longgoal_id)){
-                                        dellong.push(result[i].id)
-                                    }
+                if (difference.length != 0) {
+                    conn.queryAsync(`SELECT * FROM longgoal JOIN middlegoal on longgoal.longgoal_id = middlegoal.longgoal_id;`)
+                        .then((result) => {
+                            console.log(9999999999999)
+                            console.log(result)
+                            var dellong = new Array();
+                            for (i in result) {
+                                if (difference.includes(result[i].longgoal_id)) {
+                                    dellong.push(result[i].id)
                                 }
-                                middlegoaldeldata=`(${dellong})`
-                                console.log(middlegoaldeldata);
-                                conn.queryAsync(`SELECT * FROM middlegoal JOIN shortgoal on middlegoal.id = shortgoal.middelgoal_id;`)
-                                .then(shorres=>{
-                                    console.log('shorres',shorres)
+                            }
+                            middlegoaldeldata = `(${dellong})`
+                            console.log(middlegoaldeldata);
+                            conn.queryAsync(`SELECT * FROM middlegoal JOIN shortgoal on middlegoal.id = shortgoal.middelgoal_id;`)
+                                .then(shorres => {
+                                    console.log('shorres', shorres)
                                     var delmid = new Array();
                                     var delmidname = new Array();
-                                    for(i in shorres){
-                                        if(dellong.includes(shorres[i].middelgoal_id)){
+                                    for (i in shorres) {
+                                        if (dellong.includes(shorres[i].middelgoal_id)) {
                                             delmid.push(shorres[i].middlegoal_id)
                                             delmidname.push((shorres[i].shortgoal).replace(/(\s*)/g, ""))
                                         }
                                     }
                                     shortgoaldeldata = `(${delmid})`
                                     deltable = `${delmidname}`
-                                    console.log('deldata:',deldata);
-                                    console.log('middlegoaldeldata:',middlegoaldeldata);
-                                    console.log('shortgoaldeldata:',shortgoaldeldata);
+                                    console.log('deldata:', deldata);
+                                    console.log('middlegoaldeldata:', middlegoaldeldata);
+                                    console.log('shortgoaldeldata:', shortgoaldeldata);
                                     console.log(deltable);
-                                    var query = 
-                                    `DELETE FROM longgoal WHERE longgoal_id in ${deldata};`+
-                                    `DELETE FROM middlegoal WHERE longgoal_id in ${deldata};`+
-                                    `DELETE FROM shortgoal WHERE middelgoal_id in ${middlegoaldeldata};`+
-                                    `DROP TABLES ${deltable};`;
+                                    var query =
+                                        `DELETE FROM longgoal WHERE longgoal_id in ${deldata};` +
+                                        `DELETE FROM middlegoal WHERE longgoal_id in ${deldata};` +
+                                        `DELETE FROM shortgoal WHERE middelgoal_id in ${middlegoaldeldata};` +
+                                        `DROP TABLES ${deltable};`;
                                     console.log(query)
-                                    conn.queryAsync(query).then(()=>{
+                                    conn.queryAsync(query).then(() => {
                                         console.log('sucesssssssssssssss')
                                     })
                                 })
-                            });
+                        });
+                }
+
 
 
                 for (i in ret) {
@@ -248,17 +251,17 @@ module.exports = {
                     }
                 }
                 console.log(chagedata);
-                console.log('shortgoallafasf : ',ret)
+                console.log('shortgoallafasf : ', ret)
                 console.log(req.body)
 
                 function updatelonggoal(shorgoal_m, callback) {
-                    console.log("shortgoal_m : ",shorgoal_m,Object.values(shorgoal_m)[0],Object.keys(shorgoal_m)[0])
+                    console.log("shortgoal_m : ", shorgoal_m, Object.values(shorgoal_m)[0], Object.keys(shorgoal_m)[0])
                     conn.queryAsync(`UPDATE shortgoal SET shortgoal = '${Object.values(shorgoal_m)[0]}' where id = ${Object.keys(shorgoal_m)[0]}`)
                         .then(ret => {
                             var shortgoal_1 = Object.values(shorgoal_m)[1].replace(/(\s*)/g, "")
                             var shortgoal_2 = Object.values(shorgoal_m)[0].replace(/(\s*)/g, "")
                             conn.queryAsync(`RENAME TABLE ${shortgoal_1} TO ${shortgoal_2}`)
-                            .catch(err=>{console.log(err)})
+                                .catch(err => { console.log(err) })
                             callback(null)
                         }).catch(err => {
                             console.log(err);
@@ -294,7 +297,7 @@ module.exports = {
                         });
                 }
             });
-            
+
         })
 
     },
@@ -304,108 +307,108 @@ module.exports = {
         console.log(req.body)
         console.log(100)
         Promise.using(pool_1.connect(), conn => {
-                    console.log(100000000000);
-                    conn.queryAsync(`SELECT * FROM ${shortgoal};`).then((ret) => {
-                        console.log(ret);
-                        return ret
-                    }).then(ret => {
-                        console.log(ret)
-                        //데이터 삭제하기
-                        var pkey = []
-                        var rkey = Object.keys(req.body)
-                        var rkeys = new Array();
-                        for (i in rkey) {
-                            rkeys.push(parseInt(rkey[i]))
-                        }
-                        for (i in Object.keys(ret)) {
-                            pkey.push(ret[i].id)
-                        }
-                        console.log(req.body)
-                        console.log('rkeys: ', rkeys);
-                        console.log('pkey:', pkey);
-                        let difference = pkey.filter(x => !rkeys.includes(x));
-                        console.log('difference : ', difference);
-                        var deldata = `(${difference})`
-                        console.log(deldata)
-                        if (deldata.length != 2) {
-                            conn.queryAsync(`DELETE FROM ${shortgoal} WHERE id in ${deldata}`)
-                                .then(delresult => {
+            console.log(100000000000);
+            conn.queryAsync(`SELECT * FROM ${shortgoal};`).then((ret) => {
+                console.log(ret);
+                return ret
+            }).then(ret => {
+                console.log(ret)
+                //데이터 삭제하기
+                var pkey = []
+                var rkey = Object.keys(req.body)
+                var rkeys = new Array();
+                for (i in rkey) {
+                    rkeys.push(parseInt(rkey[i]))
+                }
+                for (i in Object.keys(ret)) {
+                    pkey.push(ret[i].id)
+                }
+                console.log(req.body)
+                console.log('rkeys: ', rkeys);
+                console.log('pkey:', pkey);
+                let difference = pkey.filter(x => !rkeys.includes(x));
+                console.log('difference : ', difference);
+                var deldata = `(${difference})`
+                console.log(deldata)
+                if (deldata.length != 2) {
+                    conn.queryAsync(`DELETE FROM ${shortgoal} WHERE id in ${deldata}`)
+                        .then(delresult => {
 
-                                });
-                            console.log(ret);
-                            console.log(req.body);
-                        }
-                        // 데이터 변한곳
-                        for (i in ret) {
-                            var data = new Object()
-                            data[ret[i].id]=new Array(ret[i].do,ret[i].date)
-                            if (req.body[ret[i].id][1] != ret[i].do) {
-                                data[ret[i].id] = req.body[ret[i].id]
-                            }
-                            var m
-                            var day
-                            var date = ret[i].date
-                            if (parseInt(date.getMonth()) < 9) {
-                                m = '0' + (date.getMonth() + 1)
-                            } else {
-                                m = date.getMonth() + 1
-                            }
-        
-                            if (parseInt(date.getDate()) < 9) {
-                                day = `0${date.getDate()}`
-                            } else {
-                                day = date.getDate()
-                            }
-                            var d = `${date.getFullYear()}-${m}-${day}`
-                            if(req.body[ret[i].id][0] != d){
-                                data[ret[i].id][1]=req.body[ret[i].id][0];
-                            }
-                            chagedata.push(data)
-                        }
-                        console.log('changedat: ', chagedata);
-                        function updatelonggoal(todo, callback) {
-                            conn.queryAsync(`UPDATE ${shortgoal} SET do = '${Object.values(todo)[0][0]}', date = DATE_FORMAT('${Object.values(todo)[0][1]}',"%Y-%m-%d") where id = ${Object.keys(todo)[0]}`)
-                                .then(ret => {
-                                    console.log('datachange success')
-                                    callback(null)
-                                }).catch(err => {
-                                    callback(null)
-                                });
-                        }
+                        });
+                    console.log(ret);
+                    console.log(req.body);
+                }
+                // 데이터 변한곳
+                for (i in ret) {
+                    var data = new Object()
+                    data[ret[i].id] = new Array(ret[i].do, ret[i].date)
+                    if (req.body[ret[i].id][1] != ret[i].do) {
+                        data[ret[i].id] = req.body[ret[i].id]
+                    }
+                    var m
+                    var day
+                    var date = ret[i].date
+                    if (parseInt(date.getMonth()) < 9) {
+                        m = '0' + (date.getMonth() + 1)
+                    } else {
+                        m = date.getMonth() + 1
+                    }
 
-                        if (chagedata.length != 0) {
-                            async.each(chagedata,
-                                updatelonggoal,
-                                function (err) {
-                                    console.log(err);
-                                });
-                        }
-                        console.log(Object.keys(req.body))
-                        console.log(Object.values(req.body))
+                    if (parseInt(date.getDate()) < 9) {
+                        day = `0${date.getDate()}`
+                    } else {
+                        day = date.getDate()
+                    }
+                    var d = `${date.getFullYear()}-${m}-${day}`
+                    if (req.body[ret[i].id][0] != d) {
+                        data[ret[i].id][1] = req.body[ret[i].id][0];
+                    }
+                    chagedata.push(data)
+                }
+                console.log('changedat: ', chagedata);
+                function updatelonggoal(todo, callback) {
+                    conn.queryAsync(`UPDATE ${shortgoal} SET do = '${Object.values(todo)[0][0]}', date = DATE_FORMAT('${Object.values(todo)[0][1]}',"%Y-%m-%d") where id = ${Object.keys(todo)[0]}`)
+                        .then(ret => {
+                            console.log('datachange success')
+                            callback(null)
+                        }).catch(err => {
+                            callback(null)
+                        });
+                }
 
-                        //추가하는 부분
-                        let pdifference = rkeys.filter(x => !pkey.includes(x));
-                        console.log('pdifference : ', pdifference);
-                        var plusdata = ``
-                        for (i in pdifference) {
-                            console.log(req.body[pdifference[i]])
-                            plusdata += `('${req.body[pdifference[i]][0]}',' ${req.body[pdifference[i]][1]}'),`
-                        }
-                        console.log('plusdata:', plusdata)
-                        console.log(`INSERT INTO ${shortgoal}(date,do) VALUES` + plusdata.slice(0, -1) + `;`)
-                        if (plusdata != 0) {
-                            conn.queryAsync(`INSERT INTO ${shortgoal}(date,do) VALUES` + plusdata.slice(0, -1) + `;`)
-                                .then(ret => {
-                                    console.log(ret)
-                                }).catch(err => {
-                                    console.log(err);
-                                });
-                        }
+                if (chagedata.length != 0) {
+                    async.each(chagedata,
+                        updatelonggoal,
+                        function (err) {
+                            console.log(err);
+                        });
+                }
+                console.log(Object.keys(req.body))
+                console.log(Object.values(req.body))
 
-                        pool_1.end();
-                    }).catch((err) => {
-                        console.log(err);
-                    })
+                //추가하는 부분
+                let pdifference = rkeys.filter(x => !pkey.includes(x));
+                console.log('pdifference : ', pdifference);
+                var plusdata = ``
+                for (i in pdifference) {
+                    console.log(req.body[pdifference[i]])
+                    plusdata += `('${req.body[pdifference[i]][0]}',' ${req.body[pdifference[i]][1]}'),`
+                }
+                console.log('plusdata:', plusdata)
+                console.log(`INSERT INTO ${shortgoal}(date,do) VALUES` + plusdata.slice(0, -1) + `;`)
+                if (plusdata != 0) {
+                    conn.queryAsync(`INSERT INTO ${shortgoal}(date,do) VALUES` + plusdata.slice(0, -1) + `;`)
+                        .then(ret => {
+                            console.log(ret)
+                        }).catch(err => {
+                            console.log(err);
+                        });
+                }
+
+                pool_1.end();
+            }).catch((err) => {
+                console.log(err);
+            })
         })
 
     },
@@ -491,5 +494,5 @@ module.exports = {
         })
 
     }
-    
+
 }
